changed NaivePrecomputed bitmap to be stack-allocated; changed pageRanks vector to use ushort

+++ Code/Simple/PreallocatedPrecomputed/Node.h
@@ -20,7 +20,7 @@
 #include <assert.h>
 
 typedef std::vector<bool> bitmap_t;
-typedef std::vector<uint> blockRanksVector;
+typedef std::vector<ushort> blockRanksVector;
 
 using namespace std;
